♻️ Refactor Login page to Server component
